returnData passing via gelatoMultiSend implemented
wrote new Fee logic

implemented Global State and ProviderStateSetter contracts

removed internalgas require and implemented safeMath check

updated tests

fixed executor underflow in providerPayables

added ProviderStateSetterFactory

renamed providerStateSetter

pre return data passing

created GelatoMultiSend

first version of GelatoMultiSend implemented

tests pass

implemented fee model

tests are passing

tested withdraw from batch exchange

first returnData memoroy passing version functional

+++ contracts/dapp_interfaces/gnosis/IBatchExchange.sol
@@ -17,7 +17,7 @@ interface IBatchExchange {
     function withdraw(address user, address token)
         external;
 
-    function deposit(address token, uint256 amount)
+    function deposit(address token, uint128 amount)
         external;
 
     function getPendingWithdraw(address user, address token)
